backend/requirements.txt
fastapi
uvicorn
newspaper3k
transformers
torch
lxml
lxml_html_clean